Improve structure of the tests
This makes bare "pytest" work for running the tests.

The test settings are now in a separate file, so they are more cleanly
separated (and not tied to a runner script).

There is also some unrelated fixes to tox.ini.

There is no real need for a wrapper script anymore, but a bigger project
might need one later, so therefore I added a bare runtests script that
runs tox.  More functionality can be added to that later, if something
is required.